backend/requirements.txt
flask
flask-cors
tensorflow>=2.18.0
numpy
pillow
groq
python-dotenv